Struct teaclave_proto::teaclave_management_service_proto::teaclave_management_client::TeaclaveManagementClient

pub struct TeaclaveManagementClient<T> { /* private fields */ }

Implementations

impl<T> TeaclaveManagementClient<T>where
T: GrpcService<BoxBody>,
T::Error: Into<StdError>,
T::ResponseBody: Body<Data = Bytes> + Send + 'static,
<T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new_with_builtin_config(inner: T) -> Self

impl TeaclaveManagementClient<Channel>

pub async fn connect<D>(dst: D) -> Result<Self, Error>where
D: TryInto<Endpoint>,
D::Error: Into<StdError>,

Attempt to create a new client by connecting to a given endpoint.

impl<T> TeaclaveManagementClient<T>where
T: GrpcService<BoxBody>,
T::Error: Into<StdError>,
T::ResponseBody: Body<Data = Bytes> + Send + 'static,
<T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_origin(inner: T, origin: Uri) -> Self

pub fn with_interceptor<F>(
inner: T,
interceptor: F
) -> TeaclaveManagementClient<InterceptedService<T, F>>where
F: Interceptor,
T::ResponseBody: Default,
T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>,
<T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_compressed(self, encoding: CompressionEncoding) -> Self

Compress requests with the given encoding.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_compressed(self, encoding: CompressionEncoding) -> Self

Enable decompressing responses.

pub fn max_decoding_message_size(self, limit: usize) -> Self

Limits the maximum size of a decoded message.

Default: 4MB

pub fn max_encoding_message_size(self, limit: usize) -> Self

Limits the maximum size of an encoded message.

Default: usize::MAX

pub async fn register_input_file(
&mut self,
request: impl IntoRequest<RegisterInputFileRequest>
) -> Result<Response<RegisterInputFileResponse>, Status>

pub async fn register_output_file(
&mut self,
request: impl IntoRequest<RegisterOutputFileRequest>
) -> Result<Response<RegisterOutputFileResponse>, Status>

pub async fn update_input_file(
&mut self,
request: impl IntoRequest<UpdateInputFileRequest>
) -> Result<Response<UpdateInputFileResponse>, Status>

pub async fn update_output_file(
&mut self,
request: impl IntoRequest<UpdateOutputFileRequest>
) -> Result<Response<UpdateOutputFileResponse>, Status>

pub async fn register_fusion_output(
&mut self,
request: impl IntoRequest<RegisterFusionOutputRequest>
) -> Result<Response<RegisterFusionOutputResponse>, Status>

pub async fn register_input_from_output(
&mut self,
request: impl IntoRequest<RegisterInputFromOutputRequest>
) -> Result<Response<RegisterInputFromOutputResponse>, Status>

pub async fn get_output_file(
&mut self,
request: impl IntoRequest<GetOutputFileRequest>
) -> Result<Response<GetOutputFileResponse>, Status>

pub async fn get_input_file(
&mut self,
request: impl IntoRequest<GetInputFileRequest>
) -> Result<Response<GetInputFileResponse>, Status>

pub async fn register_function(
&mut self,
request: impl IntoRequest<RegisterFunctionRequest>
) -> Result<Response<RegisterFunctionResponse>, Status>

pub async fn update_function(
&mut self,
request: impl IntoRequest<UpdateFunctionRequest>
) -> Result<Response<UpdateFunctionResponse>, Status>

pub async fn get_function(
&mut self,
request: impl IntoRequest<GetFunctionRequest>
) -> Result<Response<GetFunctionResponse>, Status>

pub async fn get_function_usage_stats(
&mut self,
request: impl IntoRequest<GetFunctionUsageStatsRequest>
) -> Result<Response<GetFunctionUsageStatsResponse>, Status>

pub async fn delete_function(
&mut self,
request: impl IntoRequest<DeleteFunctionRequest>
) -> Result<Response<()>, Status>

pub async fn disable_function(
&mut self,
request: impl IntoRequest<DisableFunctionRequest>
) -> Result<Response<()>, Status>

pub async fn list_functions(
&mut self,
request: impl IntoRequest<ListFunctionsRequest>
) -> Result<Response<ListFunctionsResponse>, Status>

pub async fn create_task(
&mut self,
request: impl IntoRequest<CreateTaskRequest>
) -> Result<Response<CreateTaskResponse>, Status>

pub async fn get_task(
&mut self,
request: impl IntoRequest<GetTaskRequest>
) -> Result<Response<GetTaskResponse>, Status>

pub async fn assign_data(
&mut self,
request: impl IntoRequest<AssignDataRequest>
) -> Result<Response<()>, Status>

pub async fn approve_task(
&mut self,
request: impl IntoRequest<ApproveTaskRequest>
) -> Result<Response<()>, Status>

pub async fn invoke_task(
&mut self,
request: impl IntoRequest<InvokeTaskRequest>
) -> Result<Response<()>, Status>

pub async fn cancel_task(
&mut self,
request: impl IntoRequest<CancelTaskRequest>
) -> Result<Response<()>, Status>

pub async fn save_logs(
&mut self,
request: impl IntoRequest<SaveLogsRequest>
) -> Result<Response<()>, Status>

pub async fn query_audit_logs(
&mut self,
request: impl IntoRequest<QueryAuditLogsRequest>
) -> Result<Response<QueryAuditLogsResponse>, Status>