Struct teaclave_proto::teaclave_storage_service_proto::teaclave_storage_client::TeaclaveStorageClient

pub struct TeaclaveStorageClient<T> { /* private fields */ }

Implementations

impl<T> TeaclaveStorageClient<T>where
T: GrpcService<BoxBody>,
T::Error: Into<StdError>,
T::ResponseBody: Body<Data = Bytes> + Send + 'static,
<T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new_with_builtin_config(inner: T) -> Self

impl TeaclaveStorageClient<Channel>

pub async fn connect<D>(dst: D) -> Result<Self, Error>where
D: TryInto<Endpoint>,
D::Error: Into<StdError>,

Attempt to create a new client by connecting to a given endpoint.

impl<T> TeaclaveStorageClient<T>where
T: GrpcService<BoxBody>,
T::Error: Into<StdError>,
T::ResponseBody: Body<Data = Bytes> + Send + 'static,
<T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_origin(inner: T, origin: Uri) -> Self

pub fn with_interceptor<F>(
inner: T,
interceptor: F
) -> TeaclaveStorageClient<InterceptedService<T, F>>where
F: Interceptor,
T::ResponseBody: Default,
T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>,
<T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_compressed(self, encoding: CompressionEncoding) -> Self

Compress requests with the given encoding.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_compressed(self, encoding: CompressionEncoding) -> Self

Enable decompressing responses.

pub fn max_decoding_message_size(self, limit: usize) -> Self

Limits the maximum size of a decoded message.

Default: 4MB

pub fn max_encoding_message_size(self, limit: usize) -> Self

Limits the maximum size of an encoded message.

Default: usize::MAX

pub async fn get(
&mut self,
request: impl IntoRequest<GetRequest>
) -> Result<Response<GetResponse>, Status>

pub async fn put(
&mut self,
request: impl IntoRequest<PutRequest>
) -> Result<Response<()>, Status>

pub async fn delete(
&mut self,
request: impl IntoRequest<DeleteRequest>
) -> Result<Response<()>, Status>

pub async fn enqueue(
&mut self,
request: impl IntoRequest<EnqueueRequest>
) -> Result<Response<()>, Status>

pub async fn dequeue(
&mut self,
request: impl IntoRequest<DequeueRequest>
) -> Result<Response<DequeueResponse>, Status>

pub async fn get_keys_by_prefix(
&mut self,
request: impl IntoRequest<GetKeysByPrefixRequest>
) -> Result<Response<GetKeysByPrefixResponse>, Status>